update: add isSafe check test and rename file
